backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
python-multipart==0.0.9
openpyxl==3.1.2
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.10.4
google-generativeai>=0.8.0
aiofiles==23.2.1
